Move common utils inside AbstractStatementParameterProcessor class
