+++ native/src/main/java/io/ballerina/stdlib/sql/parameterprocessor/AbstractStatementParameterProcessor.java
@@ -18,30 +18,34 @@
 
 package io.ballerina.stdlib.sql.parameterprocessor;
 
-import io.ballerina.runtime.api.types.Field;
+import io.ballerina.runtime.api.TypeTags;
+import io.ballerina.runtime.api.values.BArray;
+import io.ballerina.runtime.api.values.BDecimal;
 import io.ballerina.runtime.api.values.BObject;
+import io.ballerina.runtime.api.values.BString;
+import io.ballerina.runtime.api.values.BXml;
+import io.ballerina.stdlib.sql.Constants;
 import io.ballerina.stdlib.sql.exception.ApplicationError;
 
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import java.sql.Types;
 
 /**
- * This class has abstract implementation of methods required convert ballerina types into SQL types and
- * other methods that process the parameters of the statement.
+ * This class has abstract implementation of methods required process the JDBC Prepared Statement.
+ * Setters are used to set designated parameters to the PreparedStatement.
  *
  * @since 0.5.6
  */
-abstract class AbstractStatementParameterProcessor {
+public abstract class AbstractStatementParameterProcessor {
 
-    protected abstract int getCustomOutParameterType(BObject typedValue) throws ApplicationError;
+    public abstract int getCustomOutParameterType(BObject typedValue) throws ApplicationError;
     protected abstract int getCustomSQLType(BObject typedValue) throws ApplicationError;
+
     protected abstract void setCustomSqlTypedParam(Connection connection, PreparedStatement preparedStatement,
             int index, BObject typedValue) throws SQLException, ApplicationError, IOException;
-    protected abstract Object[] getCustomArrayData(Object value) throws ApplicationError;
-    protected abstract Object[] getCustomStructData(Connection conn, Object value) 
-            throws SQLException, ApplicationError;
     protected abstract void setVarchar(PreparedStatement preparedStatement, int index, Object value)
             throws SQLException;
     protected abstract void setVarcharArray(Connection conn, PreparedStatement preparedStatement, int index,
@@ -133,16 +137,334 @@ abstract class AbstractStatementParameterProcessor {
             throws SQLException, ApplicationError;
     protected abstract void setTimeArray(Connection conn, PreparedStatement preparedStatement, int index,
                                               Object value) throws SQLException, ApplicationError;
-    protected abstract Object[] getIntArrayData(Object value) throws ApplicationError;
-    protected abstract Object[] getFloatArrayData(Object value) throws ApplicationError;
-    protected abstract Object[] getDecimalArrayData(Object value) throws ApplicationError;
-    protected abstract Object[] getStringArrayData(Object value) throws ApplicationError;
-    protected abstract Object[] getBooleanArrayData(Object value) throws ApplicationError;
-    protected abstract Object[] getNestedArrayData(Object value) throws ApplicationError;
+    protected abstract void setXml(PreparedStatement preparedStatement, int index, BXml value) throws SQLException;
 
-    protected abstract void getRecordStructData(Connection conn, Object[] structData, int i, Object bValue)
-            throws SQLException, ApplicationError;
-    protected abstract void getArrayStructData(Field field, Object[] structData, String structuredSQLType, int i,
-            Object bValue) throws SQLException, ApplicationError;
+    public void setParams(Connection connection, PreparedStatement preparedStatement, BObject paramString)
+            throws SQLException, ApplicationError, IOException {
+        BArray arrayValue = paramString.getArrayValue(Constants.ParameterizedQueryFields.INSERTIONS);
+        for (int i = 0; i < arrayValue.size(); i++) {
+            Object object = arrayValue.get(i);
+            int index = i + 1;
+            setSQLValueParam(connection, preparedStatement, index, object, false);
+        }
+    }
+
+    public int setSQLValueParam(Connection connection, PreparedStatement preparedStatement, int index, Object object,
+                                boolean returnType) throws SQLException, ApplicationError, IOException {
+        if (object == null) {
+            preparedStatement.setNull(index, Types.NULL);
+            return Types.NULL;
+        } else if (object instanceof BString) {
+            preparedStatement.setString(index, object.toString());
+            return Types.VARCHAR;
+        } else if (object instanceof Long) {
+            preparedStatement.setLong(index, (Long) object);
+            return Types.BIGINT;
+        } else if (object instanceof Double) {
+            preparedStatement.setDouble(index, (Double) object);
+            return Types.DOUBLE;
+        } else if (object instanceof BDecimal) {
+            preparedStatement.setBigDecimal(index, ((BDecimal) object).decimalValue());
+            return Types.NUMERIC;
+        } else if (object instanceof Boolean) {
+            preparedStatement.setBoolean(index, (Boolean) object);
+            return Types.BOOLEAN;
+        } else if (object instanceof BArray) {
+            BArray objectArray = (BArray) object;
+            String type = objectArray.getElementType().toString();
+            if (objectArray.getElementType().getTag() == TypeTags.BYTE_TAG) {
+                preparedStatement.setBytes(index, objectArray.getBytes());
+                return Types.VARBINARY;
+            } else if (objectArray.getElementType().getTag() == TypeTags.ARRAY_TAG ||
+                    type.equals(Constants.SqlTypes.OPTIONAL_BYTE) || type.equals(Constants.SqlTypes.BYTE_ARRAY_TYPE)) {
+                setBinaryArray(connection, preparedStatement, index, objectArray);
+                return Types.VARBINARY;
+            } else if (type.equals(Constants.SqlTypes.STRING) || type.equals(Constants.SqlTypes.OPTIONAL_STRING)) {
+                setVarcharArray(connection, preparedStatement, index, objectArray);
+                return Types.VARCHAR;
+            } else if (type.equals(Constants.SqlTypes.INT) || type.equals(Constants.SqlTypes.OPTIONAL_INT)) {
+                setIntegerArray(connection, preparedStatement, index, objectArray);
+                return Types.INTEGER;
+            } else if (type.equals(Constants.SqlTypes.BOOLEAN_TYPE) ||
+                    type.equals(Constants.SqlTypes.OPTIONAL_BOOLEAN)) {
+                setBooleanArray(connection, preparedStatement, index, objectArray);
+                return Types.BOOLEAN;
+            } else if (type.equals(Constants.SqlTypes.FLOAT_TYPE) || type.equals(Constants.SqlTypes.OPTIONAL_FLOAT)) {
+                setFloatArray(connection, preparedStatement, index, objectArray);
+                return Types.FLOAT;
+            } else if (type.equals(Constants.SqlTypes.DECIMAL_TYPE) ||
+                    type.equals(Constants.SqlTypes.OPTIONAL_DECIMAL)) {
+                setDecimalArray(connection, preparedStatement, index, objectArray);
+                return Types.DECIMAL;
+            } else {
+                throw new ApplicationError("Invalid array type[" + type + "] set into the ParameterizedQuery.");
+            }
+        } else if (object instanceof BObject) {
+            BObject objectValue = (BObject) object;
+            if ((objectValue.getType().getTag() == TypeTags.OBJECT_TYPE_TAG)) {
+                setSqlTypedParam(connection, preparedStatement, index, objectValue);
+                if (returnType) {
+                    return getSQLType(objectValue);
+                }
+                return 0;
+            } else {
+                throw new ApplicationError("Unsupported type:" +
+                        objectValue.getType().getQualifiedName() + " in column index: " + index);
+            }
+        } else if (object instanceof BXml) {
+            setXml(preparedStatement, index, (BXml) object);
+            return Types.SQLXML;
+        } else {
+            throw new ApplicationError("Unsupported type passed in column index: " + index);
+        }
+    }
+
+    private void setSqlTypedParam(Connection connection, PreparedStatement preparedStatement, int index,
+                                  BObject typedValue)
+            throws SQLException, ApplicationError, IOException {
+        String sqlType = typedValue.getType().getName();
+        Object value = typedValue.get(Constants.TypedValueFields.VALUE);
+        switch (sqlType) {
+            case Constants.SqlTypes.VARCHAR:
+                setVarchar(preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.VARCHAR_ARRAY:
+                setVarcharArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.CHAR:
+                setChar(preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.CHAR_ARRAY:
+                setCharArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.TEXT:
+                setText(preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.NCHAR:
+                setNChar(preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.NVARCHAR_ARRAY:
+                setNVarcharArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.NVARCHAR:
+                setNVarchar(preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.BIT:
+                setBit(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.BIT_ARRAY:
+                setBitArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.BOOLEAN:
+                setBoolean(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.BOOLEAN_ARRAY:
+                setBooleanArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.INTEGER:
+                setInteger(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.INTEGER_ARRAY:
+                setIntegerArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.BIGINT:
+                setBigInt(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.BIGINT_ARRAY:
+                setBigIntArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.SMALLINT:
+                setSmallInt(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.SMALLINT_ARRAY:
+                setSmallIntArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.FLOAT:
+                setFloat(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.FLOAT_ARRAY:
+                setFloatArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.REAL:
+                setReal(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.REAL_ARRAY:
+                setRealArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.DOUBLE:
+                setDouble(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.DOUBLE_ARRAY:
+                setDoubleArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.NUMERIC:
+                setNumeric(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.NUMERIC_ARRAY:
+                setNumericArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.DECIMAL:
+                setDecimal(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.DECIMAL_ARRAY:
+                setDecimalArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.BINARY:
+                setBinary(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.BINARY_ARRAY:
+                setBinaryArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.VARBINARY:
+                setVarBinary(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.VARBINARY_ARRAY:
+                setVarBinaryArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.BLOB:
+                setBlob(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.CLOB:
+                setClob(connection, preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.NCLOB:
+                setNClob(connection, preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.DATE:
+                setDate(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.DATE_ARRAY:
+                setDateArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.TIME:
+                setTime(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.TIME_ARRAY:
+                setTimeArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.TIMESTAMP:
+                setTimestamp(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.TIMESTAMP_ARRAY:
+                setTimestampArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.DATETIME:
+                setDateTime(preparedStatement, sqlType, index, value);
+                break;
+            case Constants.SqlTypes.DATETIME_ARRAY:
+                setDateTimeArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.ARRAY:
+                setArray(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.REF:
+                setRef(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.STRUCT:
+                setStruct(connection, preparedStatement, index, value);
+                break;
+            case Constants.SqlTypes.ROW:
+                setRow(preparedStatement, sqlType, index, value);
+                break;
+            default:
+                setCustomSqlTypedParam(connection, preparedStatement, index, typedValue);
+        }
+    }
+
+    private int getSQLType(BObject typedValue) throws ApplicationError {
+        String sqlType = typedValue.getType().getName();
+        int sqlTypeValue;
+        switch (sqlType) {
+            case Constants.SqlTypes.VARCHAR:
+            case Constants.SqlTypes.TEXT:
+                sqlTypeValue = Types.VARCHAR;
+                break;
+            case Constants.SqlTypes.CHAR:
+                sqlTypeValue = Types.CHAR;
+                break;
+            case Constants.SqlTypes.NCHAR:
+                sqlTypeValue = Types.NCHAR;
+                break;
+            case Constants.SqlTypes.NVARCHAR:
+                sqlTypeValue = Types.NVARCHAR;
+                break;
+            case Constants.SqlTypes.BIT:
+                sqlTypeValue = Types.BIT;
+                break;
+            case Constants.SqlTypes.BOOLEAN:
+                sqlTypeValue = Types.BOOLEAN;
+                break;
+            case Constants.SqlTypes.INTEGER:
+                sqlTypeValue = Types.INTEGER;
+                break;
+            case Constants.SqlTypes.BIGINT:
+                sqlTypeValue = Types.BIGINT;
+                break;
+            case Constants.SqlTypes.SMALLINT:
+                sqlTypeValue = Types.SMALLINT;
+                break;
+            case Constants.SqlTypes.FLOAT:
+                sqlTypeValue = Types.FLOAT;
+                break;
+            case Constants.SqlTypes.REAL:
+                sqlTypeValue = Types.REAL;
+                break;
+            case Constants.SqlTypes.DOUBLE:
+                sqlTypeValue = Types.DOUBLE;
+                break;
+            case Constants.SqlTypes.NUMERIC:
+                sqlTypeValue = Types.NUMERIC;
+                break;
+            case Constants.SqlTypes.DECIMAL:
+                sqlTypeValue = Types.DECIMAL;
+                break;
+            case Constants.SqlTypes.BINARY:
+                sqlTypeValue = Types.BINARY;
+                break;
+            case Constants.SqlTypes.VARBINARY:
+                sqlTypeValue = Types.VARBINARY;
+                break;
+            case Constants.SqlTypes.BLOB:
+                if (typedValue instanceof BArray) {
+                    sqlTypeValue = Types.VARBINARY;
+                } else {
+                    sqlTypeValue = Types.LONGVARBINARY;
+                }
+                break;
+            case Constants.SqlTypes.CLOB:
+            case Constants.SqlTypes.NCLOB:
+                if (typedValue instanceof BString) {
+                    sqlTypeValue = Types.CLOB;
+                } else {
+                    sqlTypeValue = Types.LONGVARCHAR;
+                }
+                break;
+            case Constants.SqlTypes.DATE:
+                sqlTypeValue = Types.DATE;
+                break;
+            case Constants.SqlTypes.TIME:
+                sqlTypeValue = Types.TIME;
+                break;
+            case Constants.SqlTypes.TIMESTAMP:
+            case Constants.SqlTypes.DATETIME:
+                sqlTypeValue = Types.TIMESTAMP;
+                break;
+            case Constants.SqlTypes.ARRAY:
+                sqlTypeValue = Types.ARRAY;
+                break;
+            case Constants.SqlTypes.REF:
+                sqlTypeValue = Types.REF;
+                break;
+            case Constants.SqlTypes.STRUCT:
+                sqlTypeValue = Types.STRUCT;
+                break;
+            case Constants.SqlTypes.ROW:
+                sqlTypeValue = Types.ROWID;
+                break;
+            default:
+                sqlTypeValue = getCustomSQLType(typedValue);
+        }
+        return sqlTypeValue;
+    }
 
 }
